ID-365 Move default feature detection set of rules into its own file so it can be imported from other projects; remove jQuery dependency from feature-detect code
